requirements/base.txt
Django==1.8
django-extensions==1.5.5
django-model-utils==2.2
djangorestframework==3.1.1
ipython==3.1.0
psycopg2==2.6
whitenoise==1.0.6
